pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dosedynamics"
version = "0.1.0"
description = "Config-driven DLC analysis pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "matplotlib",
  "scipy",
  "pyyaml",
  "pydantic",
  "tensorly",
  "opencv-python",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "pre-commit",
]

[project.scripts]
dosedynamics = "dosedynamics.cli:main"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]